index , insert compelte

+++ src/main/resources/schema.sql
@@ -11,9 +11,9 @@ CREATE TABLE t_zhangzu (
     id IDENTITY NOT NULL PRIMARY KEY,
     z_date VARCHAR(255),
     z_name VARCHAR(255),
-    z_amount DOUBLE,
+    z_amount BIGINT,
     z_type VARCHAR(255),
     z_io_div VARCHAR(255),
     z_remark VARCHAR(255),
-    z_m_amount DOUBLE
+    z_m_amount BIGINT
 );
